EIRatio Edit
I removed DavidEIRatio from the branch since that would be more confusing. Instead, made the EIRatio code compatible with extra functionality that DavidEIRatio had. It should behave normally, but it now requires one more boolean argument called "filtered" which triggers the functionality of running EIRatio with selected units. This relies on the function UnitsCompare, which is also added in this commit